Roughest of drafts
Will add back-end eventually as a "mock" local/offline SQL server, maybe Azure. Also will add the data classes for specific fields of patient health (see them all listed in multi-line comment in Patient.cs as member variables) as I have content to add to them. Ultimately will add a front-end as a desktop or web app and that's why I went with C# but subject to change; could be easily translated over to Java or C++ in a few minutes.